refactor(telegram): extract sequential key module
(cherry picked from commit 19291511030769f6210abe82bad81e842177ac5d)
